django_backend/requirements.txt
# Django Backend Requirements for dirReactFinal Migration
# 2025-01-27: Initial setup for Django backend migration

Django==5.0.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
django-extensions
django-redis
django-celery-beat
celery==5.3.4
psycopg2-binary==2.9.9
Pillow==10.1.0
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
